+++ flutter_plugin_annotations/test/method_call_plugin_test.dart
@@ -1,0 +1,12 @@
+import 'package:flutter_plugin_annotations/flutter_plugin_annotations.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+
+void main() {
+  group('Testing the creation of the MethodCallPlugin class', () {
+    test('Creating an MethodCallPlugin with channelNamel', () {
+      final plugin = MethodCallPlugin(channelName: 'some name');
+      expect(plugin.channelName, 'some name');
+    });
+  });
+}