+++ flutter_plugin_annotations/CHANGELOG.md
@@ -1,1 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] 
+- Made channelName required.
+- Included documentation.
+- Included simple test.
+
+## [0.0.1] 
+
+- Initial release.